third_party/WebKit/LayoutTests/svg/animations/getCurrentTime-pause-unpause-setCurrentTime-expected.txt

PASS svg.getCurrentTime() is 0.5
PASS svg.getCurrentTime() is 1
PASS svg.getCurrentTime() is 2
PASS svg.getCurrentTime() is 3.5
